internal: split api.lua to logical blocks
The patch splits a large file `api.lua` with a Cartridge
role for a few smaller files based on their functionality.
It will help to create a Tarantool 3.0 role from that blocks
instead of copy-paste.

Part of #68
