+++ sharded-queue-scm-1.rockspec
@@ -30,6 +30,10 @@ build = {
             ['sharded_queue.stash'] = 'sharded_queue/stash.lua',
             ['sharded_queue.state'] = 'sharded_queue/state.lua',
             ['sharded_queue.stats.storage'] = 'sharded_queue/stats/storage.lua',
+            ['sharded_queue.router.config'] = 'sharded_queue/router/config.lua',
+            ['sharded_queue.router.metrics'] = 'sharded_queue/router/metrics.lua',
+            ['sharded_queue.router.queue'] = 'sharded_queue/router/queue.lua',
+            ['sharded_queue.router.tube'] = 'sharded_queue/router/tube.lua',
             ['sharded_queue.storage.config'] = 'sharded_queue/storage/config.lua',
             ['sharded_queue.storage.drivers'] = 'sharded_queue/storage/drivers.lua',
             ['sharded_queue.storage.methods'] = 'sharded_queue/storage/methods.lua',
